// src/lib/domain/database-type.ts
export enum DatabaseType {
  GENERIC = 'generic',
  POSTGRESQL = 'postgresql',
  MYSQL = 'mysql',
  SQL_SERVER = 'sql_server',
  MARIADB = 'mariadb',
  SQLITE = 'sqlite',
  CLICKHOUSE = 'clickhouse',
  COCKROACHDB = 'cockroachdb',
  ORACLE = 'oracle',
}

export const databaseTypesWithCommentSupport: DatabaseType[] = [
  DatabaseType.POSTGRESQL,
  DatabaseType.COCKROACHDB,
  DatabaseType.ORACLE,
];